requirements.txt
pandas
matplotlib
speedtest-cli